move the get*() functions to assetload (prepare to merge)
