+++ main.py
@@ -11,99 +11,15 @@ from PIL import Image
 from nextcord.ext import commands
 from lang import cmds, keywords
 import collections
+from assetload import idtoblock,blockinfos,locale,init
 
 intents = nextcord.Intents.default()
 intents.members = True
 intents.message_content = True
 bot = commands.Bot(command_prefix="!", intents=intents, help_command=None)
 print(cmds.ping.desc)
-idtoblock = {}
 
-blockinfos = collections.defaultdict(dict)
-
-locale={}
-
-localedir='localization'
-lang='english'
-
-def capitalize(s):
-    s=s[0].upper()+s[1:]
-    return s
-
-def plural(s):
-    if s.endswith('y'): # just in case
-        s=s[:-1]+'ie'
-    s+="s"
-    return s
-
-def past(s):
-    if s.endswith('e'):
-        s=s[:-1]
-    s+="ed"
-    return s
-
-modifiers={
-    '^':capitalize,
-    's':plural,
-    'd':past,
-}
-
-def getblockids():
-    with open("block_id_.smp") as f:
-        data=smp.getsmpvalue(f.read())
-    for name,i in data.items():
-        blockinfos[name]["id"] = int(i)
-        idtoblock[int(i)] = name
-
-def geticoncoords():
-    with open("block_icons.smp") as f:
-        data=smp.getsmpvalue(f.read())
-    for icon,xy in data.items():
-        x,y=xy.split(',')
-        blockinfos[icon]["iconcoord"] = (int(x), int(y))
-
-def substitutelocale(s):
-    i=0 # the index to look for the next opening bracket at
-    while '{' in s[i:]:
-        i1=s.index('{',i)
-        i2=s.find('}',i1)
-        if i2==-1: # there is no closing bracket
-            break
-        p1=s[:i1]
-        p2=s[i1+1:i2]
-        p3=s[i2+1:]
-        i=i1+1
-        if '|' in p2:
-            p2,modifier=p2.split('|',maxsplit=1)
-        else:
-            modifier=''
-        key=tuple(part.strip() for part in p2.split())
-        if key in locale:
-            localized=locale[key]
-            for mod in modifier:
-                localized=modifiers[mod](localized)
-            s=p1+localized+p3
-    return s
-
-def getlocale():
-    for fname in os.listdir(localedir):
-        if not fname.startswith(lang):
-            continue # skip files for a different language
-        with open(os.path.join(localedir,fname), "r") as f:
-            linesiter=iter(f)
-            for line in linesiter:
-                while line.endswith('\\\n'):
-                    line=line[:-1]+next(linesiter) # add the next line to this if this line ends with a backslash
-                line=re.sub('#.*$','',line) # remove comments
-                if '=' not in line:
-                    continue
-                key,value=line.split('=',maxsplit=1)
-                key=tuple(key.split())
-                value=value.strip()
-                locale[key]=value
-
-    for key,s in locale.items():
-        locale[key]=substitutelocale(s)
+init()
 
 def command(bot,name):
     async def _trycmd(cmd,ctx,*args,**kwargs):
